Move all uploaders under Spotlight namespace.
