--- app/uploaders/csv_uploader.rb
@@ -1,16 +1,0 @@
-# encoding: utf-8
-
-class CsvUploader < CarrierWave::Uploader::Base
-
-  # Choose what kind of storage to use for this uploader:
-  storage :file
-  # storage :fog
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
-
-
-end
