+++ app/uploaders/spotlight/csv_uploader.rb
@@ -1,0 +1,17 @@
+# encoding: utf-8
+module Spotlight
+  class CsvUploader < CarrierWave::Uploader::Base
+
+    # Choose what kind of storage to use for this uploader:
+    storage :file
+    # storage :fog
+
+    # Override the directory where uploaded files will be stored.
+    # This is a sensible default for uploaders that are meant to be mounted:
+    def store_dir
+      "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
+    end
+
+
+  end
+end
